requirements.txt
### Основные зависимости
aiohttp==3.8.6
asyncio==3.4.3
python-dotenv==1.0.0

### API клиенты
tinkoff.invest==1.0.5
moexalgo==2.3.2
requests==2.31.0

### Анализ данных
numpy==1.23.5
pandas==2.1.1
scipy==1.11.2

### Технические индикаторы
ta==0.11.0
vectorbt==0.25.4

### Логирование
loguru==0.7.2
sentry-sdk==1.34.0

### Telegram интеграция
python-telegram-bot==21.0.1
aiogram==2.25.1

### Утилиты
python-dateutil==2.8.2
pytz==2024.2
tzlocal==5.2
cachetools==5.3.0
dateparser==1.2.2
httpx==0.27.0  # Компромиссная версия